feat: move SDP munging methods to SdpUtils object
