+++ telnyx_rtc/src/main/java/com/telnyx/webrtc/sdk/TelnyxClient.kt
@@ -40,6 +40,7 @@ import com.telnyx.webrtc.lib.SessionDescription
 import java.util.*
 import kotlin.concurrent.timerTask
 import java.util.regex.Pattern
+import com.telnyx.webrtc.sdk.utilities.SdpUtils
 
 /**
  * The TelnyxClient class that can be used to control the SDK. Create / Answer calls, change audio device, etc.
@@ -231,7 +232,8 @@ class TelnyxClient(
                     updateCallState(CallState.ERROR)
                     Logger.e(message = "Failed to generate local description (Answer SDP) for call $callId")
                 } else {
-                    val finalAnswerSdp = modifyAnswerSdpToIncludeOfferCodecs(originalOfferSdp, generatedAnswerSdp)
+                    // Use the SdpUtils to modify the SDP
+                    val finalAnswerSdp = SdpUtils.modifyAnswerSdpToIncludeOfferCodecs(originalOfferSdp, generatedAnswerSdp)
                     Logger.d(tag = "SDP_Modify", message = "[Original Answer SDP]:\n$generatedAnswerSdp")
                     Logger.d(tag = "SDP_Modify", message = "[Final Answer SDP]:\n$finalAnswerSdp")
 
@@ -260,105 +262,6 @@ class TelnyxClient(
 
         this.addToCalls(acceptCall)
         return acceptCall
-    }
-
-    /**
-     * SDP Munging function to modify the generated Answer SDP to include audio codecs from the Offer SDP
-     * that might have been excluded by the WebRTC library.
-     */
-    private fun modifyAnswerSdpToIncludeOfferCodecs(offerSdp: String, answerSdp: String): String {
-        try {
-            val offerAudioCodecs = extractAudioCodecs(offerSdp)
-            if (offerAudioCodecs.isEmpty()) {
-                Logger.w(tag = "SDP_Modify", message = "No audio codecs found in Offer SDP. Returning original Answer.")
-                return answerSdp
-            }
-
-            val answerAudioCodecs = extractAudioCodecs(answerSdp)
-
-            val missingCodecs = offerAudioCodecs.filterKeys { it !in answerAudioCodecs.keys }
-            if (missingCodecs.isEmpty()) {
-                Logger.d(tag = "SDP_Modify", message = "No missing audio codecs detected. Returning original Answer.")
-                return answerSdp
-            }
-            Logger.d(tag = "SDP_Modify", message = "Missing codecs to add: ${missingCodecs.keys}")
-
-            val answerLines = answerSdp.lines().toMutableList()
-            var audioMLineIndex = -1
-            var firstAudioAttrIndex = -1
-
-            for (i in answerLines.indices) {
-                if (answerLines[i].startsWith(MEDIA_LINE_PREFIX)) {
-                    audioMLineIndex = i
-                } else if (audioMLineIndex != -1 && answerLines[i].startsWith(ATTRIBUTE_PREFIX) && firstAudioAttrIndex == -1) {
-                    firstAudioAttrIndex = i
-                } else if (audioMLineIndex != -1 && answerLines[i].startsWith(MEDIA_LINE_PREFIX)) {
-                    if (firstAudioAttrIndex == -1) firstAudioAttrIndex = i
-                    break
-                }
-            }
-
-            if (audioMLineIndex == -1 || firstAudioAttrIndex == -1) {
-                Logger.e(tag = "SDP_Modify", message = "Could not find m=audio line or attribute insertion point in Answer SDP. Returning original.")
-                return answerSdp
-            }
-
-            val originalMLine = answerLines[audioMLineIndex]
-            val mLineParts = originalMLine.split(" ").toMutableList()
-
-            // Check that the m-line has at least the minimum parts required to be valid
-            if (mLineParts.size < MINIMUM_M_LINE_PARTS) {
-                Logger.w(tag = "SDP_Modify", message = "Unexpected m=audio line format: $originalMLine. It contains less than the minimum number of parts required")
-                return answerSdp
-            }
-
-            // Extract the existing payloads
-            val existingPayloads = mLineParts.subList(PAYLOAD_START_INDEX, mLineParts.size).toSet()
-            val newPayloads = missingCodecs.keys.filter { it !in existingPayloads }
-            mLineParts.addAll(newPayloads)
-            answerLines[audioMLineIndex] = mLineParts.joinToString(" ")
-            Logger.d(tag = "SDP_Modify", message = "Modified m=audio line: ${answerLines[audioMLineIndex]}")
-
-            val rtpmapLinesToAdd = missingCodecs.values.toList()
-            answerLines.addAll(firstAudioAttrIndex, rtpmapLinesToAdd)
-            Logger.d(tag = "SDP_Modify", message = "Added rtpmap lines: $rtpmapLinesToAdd")
-
-            return answerLines.joinToString("\r\n")
-
-        } catch (e: Exception) {
-            Logger.e(tag = "SDP_Modify", message = "Error modifying SDP: ${e.message}. Returning original Answer.")
-            Logger.e(tag = "SDP_Modify", message = Log.getStackTraceString(e))
-            return answerSdp
-        }
-    }
-
-    /**
-     * Extracts audio codec payload types and their corresponding a=rtpmap lines from SDP.
-     */
-    private fun extractAudioCodecs(sdp: String): Map<String, String> {
-        val codecs = mutableMapOf<String, String>()
-        val lines = sdp.lines()
-        var inAudioSection = false
-        val rtpmapPattern = Pattern.compile("^a=rtpmap:(\\d+)\\s+(.+)$")
-
-        for (line in lines) {
-            if (line.startsWith("m=audio")) {
-                inAudioSection = true
-            } else if (line.startsWith("m=")) {
-                if (inAudioSection) break
-            }
-
-            if (inAudioSection && line.startsWith("a=rtpmap:")) {
-                val matcher = rtpmapPattern.matcher(line)
-                if (matcher.find()) {
-                    val payloadType = matcher.group(1)
-                    if (payloadType != null) {
-                        codecs[payloadType] = line
-                    }
-                }
-            }
-        }
-        return codecs
     }
 
     /**
